chore: start refactor of tests
