+++ tests/src/subspace2.rs
@@ -1,0 +1,2 @@
+mod burn;
+mod registration;
